+++ jsonc/src/main/java/house/greenhouse/greenhouseconfig/jsonc/JsonCObject.java
@@ -9,16 +9,23 @@ import java.util.*;
 
 public class JsonCObject extends JsonCElement {
 	public static final JsonCObject EMPTY = new JsonCObject(Map.of());
+	private JsonObject json = new JsonObject();
 	private Map<String, JsonCElement> members = new LinkedHashMap<>();
 
 	public JsonCObject(Map<String, JsonCElement> members) {
 		super(null);
 		this.members = members;
+		for (Map.Entry<String, JsonCElement> entry : members.entrySet()) {
+			json.add(entry.getKey(), entry.getValue().json());
+		}
 	}
 
 	public JsonCObject(Map<String, JsonCElement> members, String... comments) {
 		super(null, comments);
 		this.members = members;
+		for (Map.Entry<String, JsonCElement> entry : members.entrySet()) {
+			json.add(entry.getKey(), entry.getValue().json());
+		}
 	}
 
 	public JsonCObject(String... comments) {
@@ -27,12 +34,17 @@ public class JsonCObject extends JsonCElement {
 
 	public JsonCObject(JsonObject object, String... comments) {
 		super(null, comments);
-
 		for (Map.Entry<String, JsonElement> entry : object.entrySet()) {
 			members.put(entry.getKey(), entry.getValue().isJsonObject() ?
 					new JsonCObject(entry.getValue().getAsJsonObject()) :
 					new JsonCElement(entry.getValue()));
 		}
+		json = object;
+	}
+
+	@Override
+	public JsonElement json() {
+		return json;
 	}
 
 	public Map<String, JsonCElement> toMap() {
@@ -41,18 +53,24 @@ public class JsonCObject extends JsonCElement {
 
 	public void put(String name, JsonCElement json) {
 		members.put(name, json);
+		this.json.add(name, json.json());
 	}
 
 	public void putAll(Map<String, JsonCElement> map) {
 		members.putAll(map);
+		for (Map.Entry<String, JsonCElement> entry : map.entrySet()) {
+			json.add(entry.getKey(), entry.getValue().json());
+		}
 	}
 
 	public void remove(String name) {
 		members.remove(name);
+		json.remove(name);
 	}
 
 	public void clear() {
 		members.clear();
+		json = new JsonObject();
 	}
 
 	@Override