+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Bugfixes
+### JSONC Language
 - Fixed JSONCOps not creating a copy of the JSONC in the `remove` method.
+- Fixed JSONC objects crashing when adding to an array using JSONCOps because of no JSON .